Correção na estrutura de pastas.
