--- frontend/src/components/Common/FormControl/index.js
@@ -1,1 +1,0 @@
-export { default as FormControl } from "./FormControl";
