--- frontend/src/components/Header/UserMenu/index.js
@@ -1,1 +1,0 @@
-export { default as UserMenu } from "./UserMenu";
