--- frontend/src/components/NotFound/index.js
@@ -1,1 +1,0 @@
-export { default as NotFound } from "./NotFound";
