--- frontend/src/components/UserLogin/index.js
@@ -1,1 +1,0 @@
-export { default as UserLogin } from "./UserLogin";
