+++ frontend/src/pages/Home/Home.js
@@ -1,6 +1,6 @@
 import React from "react";
-import List from "../../components/List";
+import { List } from "../../components/List";
 
-const Home = () => <List />;
+const Home = ({ history }) => <List history={history} />;
 
 export default Home;
